refactor: Move company-related classes into aggregate package
- 회사 관련 클래스를 aggregate 패키지로 이동했습니다.
