+++ src/main/java/com/akatsuki/pioms/company/repository/CompanyRepository.java
@@ -1,7 +1,7 @@
 package com.akatsuki.pioms.company.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.akatsuki.pioms.company.entity.Company;
+import com.akatsuki.pioms.company.aggregate.Company;
 
 public interface CompanyRepository extends JpaRepository<Company, Integer> {
 }
