+++ src/main/java/com/akatsuki/pioms/company/service/CompanyService.java
@@ -1,8 +1,8 @@
 package com.akatsuki.pioms.company.service;
 
 
-import com.akatsuki.pioms.company.entity.Company;
-import com.akatsuki.pioms.company.vo.CompanyVO;
+import com.akatsuki.pioms.company.aggregate.Company;
+import com.akatsuki.pioms.company.aggregate.CompanyVO;
 
 
 public interface CompanyService {
